[test] Compile python modules
Minor speedup.

Need to distribute them from meson.build I suppose.
